Refactor: test framework
