+++ tests/test_table_client_hash.rs
@@ -1,0 +1,64 @@
+// Licensed under Apache-2.0.
+
+#[allow(unused_imports)]
+#[allow(unused)]
+mod utils;
+pub mod test_table_client_base;
+
+use obkv::{Table, Value};
+use serial_test_derive::serial;
+use test_log::test;
+
+// ```sql
+// CREATE TABLE `TEST_VARCHAR_TABLE_HASH_CONCURRENT` (
+//     `c1` bigint(20) NOT NULL,
+//     `c2` varchar(20) DEFAULT NULL,
+//     PRIMARY KEY (`c1`)
+// ) DEFAULT CHARSET = utf8mb4 COMPRESSION = 'lz4_1.0' REPLICA_NUM = 3 BLOCK_SIZE = 16384 USE_BLOOM_FILTER = FALSE TABLET_SIZE = 134217728 PCTFREE = 10
+// partition by hash(c1) partitions 16;
+// ```
+#[test]
+#[serial]
+fn test_concurrent() {
+    let client = utils::common::build_normal_client();
+    const TABLE_NAME: &str = "TEST_VARCHAR_TABLE_HASH_CONCURRENT";
+    client.add_row_key_element(TABLE_NAME, vec!["c1".to_string()]);
+    let test = test_table_client_base::BaseTest::new(client);
+
+    test.test_bigint_concurrent(TABLE_NAME);
+    test.clean_bigint_table(TABLE_NAME);
+}
+
+// ```sql
+// CREATE TABLE `TEST_TABLE_BATCH_HASH` (
+// `c1` bigint NOT NULL,
+// `c1sk` varchar(20) NOT NULL,
+// `c2` varchar(20) DEFAULT NULL,
+// PRIMARY KEY (`c1`, `c1sk`)) DEFAULT CHARSET = utf8mb4 ROW_FORMAT = DYNAMIC COMPRESSION = 'lz4_1.0' REPLICA_NUM = 3 BLOCK_SIZE = 16384 USE_BLOOM_FILTER = FALSE TABLET_SIZE = 134217728 PCTFREE = 10
+// partition by hash(`c1`) partitions 16;
+// ```
+#[test]
+fn test_obtable_client_hash() {
+    let client = utils::common::build_normal_client();
+    const TABLE_NAME: &str = "TEST_TABLE_BATCH_HASH";
+    client.add_row_key_element(TABLE_NAME, vec!["c1".to_string(), "c1sb".to_string()],);
+
+    let test_key0 : i64 = 100;
+    let test_key1 : i64 = 200;
+    // insert some data
+    let mut batch_op = client.batch_operation(4);
+    batch_op.delete(vec![Value::from(test_key0), Value::from("subKey_0")]);
+    batch_op.delete(vec![Value::from(test_key1), Value::from("subKey_1")]);
+    batch_op.insert(
+        vec![Value::from(test_key0), Value::from("subKey_0")],
+        vec!["c2".to_owned()],
+        vec![Value::from("batchValue_0")],
+    );
+    batch_op.insert(
+        vec![Value::from(test_key1), Value::from("subKey_1")],
+        vec!["c2".to_owned()],
+        vec![Value::from("batchValue_1")],
+    );
+    let result = client.execute_batch(TABLE_NAME, batch_op);
+    assert!(result.is_ok());
+}
